tensorboard/tensorboard_rest_api/requirements.txt
Flask==1.0.2
requests==2.21.0
python-dotenv==0.12.0
